+++ network.go
@@ -152,303 +152,12 @@ type JourneyCompletion struct {
 	Hops       []WorldHop `json:"hops,omitempty"` // The attestation log
 }
 
-type NewspaperEvent struct {
-	From       string
-	Status     NaraStatus
-	Signature  string // Base64-encoded signature of the status JSON
-	StatusJSON []byte `json:"-"` // Raw status JSON for signature verification
-}
-
-// SignNewspaper creates a signed newspaper event
-func (network *Network) SignNewspaper(status NaraStatus) NewspaperEvent {
-	event := NewspaperEvent{
-		From:   network.meName(),
-		Status: status,
-	}
-	// Sign the JSON-serialized status
-	statusJSON, _ := json.Marshal(status)
-	event.Signature = network.local.Keypair.SignBase64(statusJSON)
-	return event
-}
-
-// VerifyNewspaper verifies a newspaper event signature
-func (event *NewspaperEvent) Verify(publicKey []byte) bool {
-	if event.Signature == "" {
-		logrus.Warnf("Newspaper from %s is missing signature", event.From)
-		return false
-	}
-	statusJSON := event.StatusJSON
-	if len(statusJSON) == 0 {
-		var err error
-		statusJSON, err = json.Marshal(event.Status)
-		if err != nil {
-			logrus.Warnf("Failed to marshal newspaper status from %s: %v", event.From, err)
-			return false
-		}
-	}
-	return VerifySignatureBase64(publicKey, statusJSON, event.Signature)
-}
-
-type HeyThereEvent struct {
-	From      string
-	PublicKey string // Base64-encoded Ed25519 public key
-	MeshIP    string // Tailscale IP for mesh communication
-	ID        string // Nara ID: deterministic hash of soul+name
-	Signature string // Base64-encoded signature of "hey_there:{From}:{PublicKey}:{MeshIP}:{ID}"
-}
-
-// Sign signs the HeyThereEvent with the given keypair
-func (h *HeyThereEvent) Sign(kp NaraKeypair) {
-	message := h.signatureMessage()
-	h.Signature = kp.SignBase64([]byte(message))
-}
-
-// Verify verifies the HeyThereEvent signature against the embedded public key
-func (h *HeyThereEvent) Verify() bool {
-	if h.PublicKey == "" || h.Signature == "" {
-		return false
-	}
-	pubKey, err := ParsePublicKey(h.PublicKey)
-	if err != nil {
-		return false
-	}
-
-	// Try new format first (with ID)
-	message := h.signatureMessage()
-	if VerifySignatureBase64(pubKey, []byte(message), h.Signature) {
-		return true
-	}
-
-	// Fall back to old format (without ID) for backwards compatibility
-	if h.ID != "" {
-		oldMessage := fmt.Sprintf("hey_there:%s:%s:%s", h.From, h.PublicKey, h.MeshIP)
-		return VerifySignatureBase64(pubKey, []byte(oldMessage), h.Signature)
-	}
-
-	return false
-}
-
-// signatureMessage returns the message to sign based on whether ID is present
-func (h *HeyThereEvent) signatureMessage() string {
-	if h.ID != "" && false { // TODO(signature) temporarily turned off
-		return fmt.Sprintf("hey_there:%s:%s:%s:%s", h.From, h.PublicKey, h.MeshIP, h.ID)
-	}
-	// Legacy format for naras without ID
-	return fmt.Sprintf("hey_there:%s:%s:%s", h.From, h.PublicKey, h.MeshIP)
-}
-
-// ContentString implements Payload interface for HeyThereEvent
-func (h *HeyThereEvent) ContentString() string {
-	return h.signatureMessage()
-}
-
-// IsValid implements Payload interface for HeyThereEvent
-// Note: Inner signature verification is not required here - the SyncEvent signature
-// is the attestation layer. This just validates the payload has required fields.
-func (h *HeyThereEvent) IsValid() bool {
-	return h.From != "" && h.PublicKey != ""
-}
-
-// GetActor implements Payload interface for HeyThereEvent
-func (h *HeyThereEvent) GetActor() string { return h.From }
-
-// GetTarget implements Payload interface for HeyThereEvent
-func (h *HeyThereEvent) GetTarget() string { return h.From }
-
-// VerifySignature implements Payload using the embedded public key
-func (h *HeyThereEvent) VerifySignature(event *SyncEvent, lookup PublicKeyLookup) bool {
-	if h.PublicKey == "" {
-		return false
-	}
-	pubKey, err := ParsePublicKey(h.PublicKey)
-	if err != nil {
-		return false
-	}
-	return event.VerifyWithKey(pubKey)
-}
-
-// UIFormat returns UI-friendly representation
-func (h *HeyThereEvent) UIFormat() map[string]string {
-	detail := "hey there!"
-	if h.MeshIP != "" {
-		detail = fmt.Sprintf("mesh: %s", h.MeshIP)
-	}
-	return map[string]string{
-		"icon":   "👋",
-		"text":   fmt.Sprintf("%s joined the network", h.From),
-		"detail": detail,
-	}
-}
-
-// LogFormat returns technical log description
-func (h *HeyThereEvent) LogFormat() string {
-	return fmt.Sprintf("hey-there from %s (mesh: %s)", h.From, h.MeshIP)
-}
-
-// ToLogEvent returns a structured log event for the logging system
-func (h *HeyThereEvent) ToLogEvent() *LogEvent {
-	return &LogEvent{
-		Category: CategoryPresence,
-		Type:     "welcome",
-		Actor:    h.From,
-		Target:   h.From,
-		Detail:   h.LogFormat(),
-		GroupFormat: func(actors string) string {
-			return fmt.Sprintf("👋 %s said hey-there", actors)
-		},
-	}
-}
-
-type ChauEvent struct {
-	From      string
-	PublicKey string // Base64-encoded Ed25519 public key
-	ID        string // Nara ID: deterministic hash of soul+name
-	Signature string // Base64-encoded signature of "chau:{From}:{PublicKey}:{ID}"
-}
-
-// Sign signs the ChauEvent with the given keypair
-func (c *ChauEvent) Sign(kp NaraKeypair) {
-	message := c.signatureMessage()
-	c.Signature = kp.SignBase64([]byte(message))
-}
-
-// Verify verifies the ChauEvent signature against the embedded public key
-func (c *ChauEvent) Verify() bool {
-	if c.PublicKey == "" || c.Signature == "" {
-		return false
-	}
-	pubKey, err := ParsePublicKey(c.PublicKey)
-	if err != nil {
-		return false
-	}
-
-	// Try new format first (with ID)
-	message := c.signatureMessage()
-	if VerifySignatureBase64(pubKey, []byte(message), c.Signature) {
-		return true
-	}
-
-	// Fall back to old format (without ID) for backwards compatibility
-	if c.ID != "" {
-		oldMessage := fmt.Sprintf("chau:%s:%s", c.From, c.PublicKey)
-		return VerifySignatureBase64(pubKey, []byte(oldMessage), c.Signature)
-	}
-
-	return false
-}
-
-// signatureMessage returns the message to sign based on whether ID is present
-func (c *ChauEvent) signatureMessage() string {
-	if c.ID != "" && false { // TODO(signature) temporarily turned off
-		return fmt.Sprintf("chau:%s:%s:%s", c.From, c.PublicKey, c.ID)
-	}
-	// Legacy format for naras without ID
-	return fmt.Sprintf("chau:%s:%s", c.From, c.PublicKey)
-}
-
-// ContentString implements Payload interface for ChauEvent
-func (c *ChauEvent) ContentString() string {
-	return c.signatureMessage()
-}
-
-// IsValid implements Payload interface for ChauEvent
-// Note: Inner signature verification is not required here - the SyncEvent signature
-// is the attestation layer. This just validates the payload has required fields.
-func (c *ChauEvent) IsValid() bool {
-	return c.From != ""
-}
-
-// GetActor implements Payload interface for ChauEvent
-func (c *ChauEvent) GetActor() string { return c.From }
-
-// GetTarget implements Payload interface for ChauEvent
-func (c *ChauEvent) GetTarget() string { return c.From }
-
-// VerifySignature implements Payload using the embedded public key
-func (c *ChauEvent) VerifySignature(event *SyncEvent, lookup PublicKeyLookup) bool {
-	if c.PublicKey == "" {
-		return false
-	}
-	pubKey, err := ParsePublicKey(c.PublicKey)
-	if err != nil {
-		return false
-	}
-	return event.VerifyWithKey(pubKey)
-}
-
-// UIFormat returns UI-friendly representation
-func (c *ChauEvent) UIFormat() map[string]string {
-	return map[string]string{
-		"icon":   "👋",
-		"text":   fmt.Sprintf("%s left the network", c.From),
-		"detail": "chau!",
-	}
-}
-
-// LogFormat returns technical log description
-func (c *ChauEvent) LogFormat() string {
-	return fmt.Sprintf("chau from %s", c.From)
-}
-
-// ToLogEvent returns a structured log event for the logging system
-func (c *ChauEvent) ToLogEvent() *LogEvent {
-	return &LogEvent{
-		Category: CategoryPresence,
-		Type:     "goodbye",
-		Actor:    c.From,
-		Detail:   c.LogFormat(),
-		GroupFormat: func(actors string) string {
-			return fmt.Sprintf("💨 %s bounced", actors)
-		},
-	}
-}
-
 // PeerResponse contains identity information about a peer.
 // Used by the peer resolution protocol to return discovered peer info.
 type PeerResponse struct {
 	Target    string `json:"target"`
 	PublicKey string `json:"public_key"`
 	MeshIP    string `json:"mesh_ip,omitempty"`
-}
-
-// NeighborInfo contains information about a neighbor to share in howdy responses
-type NeighborInfo struct {
-	Name        string
-	PublicKey   string
-	MeshIP      string
-	ID          string          // Nara ID: deterministic hash of soul+name
-	Observation NaraObservation // What I know about this neighbor
-}
-
-// HowdyEvent is sent in response to hey_there to help with discovery and start time recovery
-type HowdyEvent struct {
-	From      string          // Who's sending this howdy
-	To        string          // Who this is in response to (hey_there sender)
-	Seq       int             // Sequence number (1-10) for coordination
-	You       NaraObservation // What I know about you (includes StartTime!)
-	Neighbors []NeighborInfo  // ~10 other naras you should know about
-	Me        NaraStatus      // My own status
-	Signature string          // Ed25519 signature
-}
-
-// Sign signs the HowdyEvent with the given keypair
-func (h *HowdyEvent) Sign(kp NaraKeypair) {
-	// Sign a deterministic representation of the event
-	message := fmt.Sprintf("howdy:%s:%s:%d", h.From, h.To, h.Seq)
-	h.Signature = kp.SignBase64([]byte(message))
-}
-
-// Verify verifies the HowdyEvent signature against the public key in Me.PublicKey
-func (h *HowdyEvent) Verify() bool {
-	if h.Me.PublicKey == "" || h.Signature == "" {
-		return false
-	}
-	pubKey, err := ParsePublicKey(h.Me.PublicKey)
-	if err != nil {
-		return false
-	}
-	message := fmt.Sprintf("howdy:%s:%s:%d", h.From, h.To, h.Seq)
-	return VerifySignatureBase64(pubKey, []byte(message), h.Signature)
 }
 
 func NewNetwork(localNara *LocalNara, host string, user string, pass string) *Network {
@@ -862,81 +571,6 @@ func (network *Network) markEmittersAsSeen(events []SyncEvent) {
 
 // processHeyThereSyncEvents extracts peer identity information from hey_there sync events.
 // This enables peer discovery through gossip without requiring MQTT broadcasts.
-func (network *Network) processHeyThereSyncEvents(events []SyncEvent) {
-	for i := range events {
-		e := &events[i]
-		if e.Service != ServiceHeyThere || e.HeyThere == nil {
-			continue
-		}
-
-		h := e.HeyThere
-		if h.From == network.meName() {
-			continue // Ignore our own hey_there events
-		}
-
-		// For hey_there events, verify using the public key FROM the payload itself.
-		// This is the bootstrap case - hey_there is how we learn public keys.
-		// The payload contains the public key, and the SyncEvent is signed with it.
-		if e.IsSigned() && h.PublicKey != "" {
-			pubKey, err := ParsePublicKey(h.PublicKey)
-			if err != nil {
-				logrus.Warnf("📡 Invalid public key in hey_there from %s: %v", h.From, err)
-				continue
-			}
-			if !e.VerifyWithKey(pubKey) {
-				logrus.Warnf("📡 Invalid hey_there SyncEvent signature from %s", h.From)
-				continue
-			}
-		}
-
-		// Check if nara exists and update or create
-		network.local.mu.Lock()
-		nara, exists := network.Neighbourhood[h.From]
-		network.local.mu.Unlock()
-
-		if exists {
-			// Update existing nara with proper locking
-			nara.mu.Lock()
-			updated := false
-			if nara.Status.PublicKey == "" && h.PublicKey != "" {
-				nara.Status.PublicKey = h.PublicKey
-				updated = true
-			}
-			if nara.Status.MeshIP == "" && h.MeshIP != "" {
-				nara.Status.MeshIP = h.MeshIP
-				nara.Status.MeshEnabled = true
-				updated = true
-			}
-			nara.mu.Unlock()
-			if updated {
-				logrus.Infof("📡 Updated identity for %s via hey_there event (🔑)", h.From)
-			}
-		} else {
-			// Create new nara and import it
-			newNara := NewNara(h.From)
-			newNara.Status.PublicKey = h.PublicKey
-			newNara.Status.MeshIP = h.MeshIP
-			newNara.Status.MeshEnabled = h.MeshIP != ""
-			network.importNara(newNara)
-			logrus.Infof("📡 Discovered new peer %s via hey_there event (🔑)", h.From)
-		}
-	}
-}
-
-// emitHeyThereSyncEvent creates and adds a hey_there sync event to our ledger.
-// This allows our identity to propagate through gossip (new mechanism replacing MQTT hey_there).
-func (network *Network) emitHeyThereSyncEvent() {
-	publicKey := FormatPublicKey(network.local.Keypair.PublicKey)
-	meshIP := network.local.Me.Status.MeshIP
-
-	event := NewHeyThereSyncEvent(network.meName(), publicKey, meshIP, network.local.ID, network.local.Keypair)
-	network.local.SyncLedger.MergeEvents([]SyncEvent{event})
-	if network.local.Projections != nil {
-		network.local.Projections.Trigger()
-	}
-	logrus.Infof("%s: 👋 (gossip)", network.meName())
-}
-
 // InitGossipIdentity initializes gossip-mode identity emission.
 // Called by Start() and can be called by tests to simulate startup.
 // This emits the hey_there sync event that allows our identity to propagate through gossip.
@@ -951,82 +585,6 @@ func (network *Network) InitGossipIdentity() {
 // This allows gossip-only naras to distinguish OFFLINE (graceful) from MISSING (timeout).
 //
 // IMPORTANT: This function directly updates observations, which can race with live events
-// during boot. The long-term direction is to remove this direct observation update and let
-// the OnlineStatus projection be the single source of truth. The observationMaintenanceOnce
-// function already syncs from projections, so removing this would unify the status logic.
-// For now, we skip processing during boot to avoid the race condition.
-func (network *Network) processChauSyncEvents(events []SyncEvent) {
-	// Skip during boot to avoid race condition where backfilled chau events
-	// can clobber ONLINE status set by live hey_there events.
-	// The projection will handle chau events correctly, and observationMaintenanceOnce
-	// will sync from projections after boot completes.
-	if network.local.isBooting() {
-		return
-	}
-
-	for i := range events {
-		e := &events[i]
-		if e.Service != ServiceChau || e.Chau == nil {
-			continue
-		}
-
-		c := e.Chau
-		if c.From == network.meName() {
-			continue // Ignore our own chau events
-		}
-
-		// Verify the SyncEvent signature (the attestation layer)
-		if !network.VerifySyncEvent(e) {
-			logrus.Warnf("📡 Invalid chau SyncEvent signature from %s", c.From)
-			continue
-		}
-
-		// Check if there's a more recent hey_there from this nara
-		// This prevents stale chau events from incorrectly marking naras offline during backfill
-		if network.hasMoreRecentHeyThere(c.From, e.Timestamp) {
-			continue
-		}
-
-		// Mark the nara as OFFLINE (graceful shutdown)
-		observation := network.local.getObservation(c.From)
-		if observation.Online == "ONLINE" {
-			observation.Online = "OFFLINE"
-			observation.LastSeen = time.Now().Unix()
-			network.local.setObservation(c.From, observation)
-			// LogService handles logging via ledger watching
-			network.Buzz.increase(2)
-		}
-	}
-}
-
-// hasMoreRecentHeyThere checks if there's a hey_there event from the given nara
-// that's more recent than the specified timestamp.
-func (network *Network) hasMoreRecentHeyThere(from string, thanTimestamp int64) bool {
-	if network.local.SyncLedger == nil {
-		return false
-	}
-
-	heyThereEvents := network.local.SyncLedger.GetEventsByService(ServiceHeyThere)
-	for _, e := range heyThereEvents {
-		if e.HeyThere != nil && e.HeyThere.From == from && e.Timestamp > thanTimestamp {
-			return true
-		}
-	}
-	return false
-}
-
-// emitChauSyncEvent creates and adds a chau sync event to our ledger.
-// This allows graceful shutdown to propagate through gossip.
-func (network *Network) emitChauSyncEvent() {
-	publicKey := FormatPublicKey(network.local.Keypair.PublicKey)
-	event := NewChauSyncEvent(network.meName(), publicKey, network.local.ID, network.local.Keypair)
-	network.local.SyncLedger.MergeEvents([]SyncEvent{event})
-	if network.local.Projections != nil {
-		network.local.Projections.Trigger()
-	}
-	// LogService handles logging via ledger watching
-}
-
 // --- Peer Resolution Protocol ---
 // When we don't know a peer's identity (public key), we can query neighbors.
 // Uses HTTP redirects: if a neighbor doesn't know, they redirect us to try someone else.
@@ -1628,257 +1186,6 @@ func (network *Network) meName() string {
 	return network.local.Me.Name
 }
 
-func (network *Network) announce() {
-	if network.ReadOnly {
-		return
-	}
-	network.testAnnounceCount++ // Track for testing
-	topic := fmt.Sprintf("%s/%s", "nara/newspaper", network.meName())
-	network.recordObservationOnlineNara(network.meName(), 0)
-
-	// Broadcast slim newspapers without Observations (observations are event-sourced)
-	network.local.Me.mu.Lock()
-	slimStatus := network.local.Me.Status
-	network.local.Me.mu.Unlock()
-	slimStatus.Observations = nil
-	if network.local.SyncLedger != nil {
-		slimStatus.EventStoreByService = network.local.SyncLedger.GetEventCountsByService()
-		slimStatus.EventStoreTotal = network.local.SyncLedger.EventCount()
-		slimStatus.EventStoreCritical = network.local.SyncLedger.GetCriticalEventCount()
-	} else {
-		slimStatus.EventStoreByService = nil
-		slimStatus.EventStoreTotal = 0
-		slimStatus.EventStoreCritical = 0
-	}
-
-	// Populate stash metrics
-	if network.stashService != nil {
-		metrics := network.stashService.GetStorageMetrics()
-		slimStatus.StashStored = metrics.StashesStored
-		slimStatus.StashBytes = metrics.TotalStashBytes
-		slimStatus.StashConfidants = network.stashService.ConfidantCount()
-	}
-
-	signedEvent := network.SignNewspaper(slimStatus)
-	network.postEvent(topic, signedEvent)
-}
-
-func (network *Network) announceForever() {
-	for {
-		// Newspapers are lightweight heartbeats (30-300s), observations are event-sourced
-		ts := network.local.chattinessRate(30, 300)
-
-		// Wait with graceful shutdown support
-		select {
-		case <-time.After(time.Duration(ts) * time.Second):
-			network.announce()
-		case <-network.ctx.Done():
-			logrus.Debugf("announceForever: shutting down gracefully")
-			return
-		}
-	}
-}
-
-func (network *Network) processNewspaperEvents() {
-	for {
-		select {
-		case event := <-network.newspaperInbox:
-			network.handleNewspaperEvent(event)
-		case <-network.ctx.Done():
-			logrus.Debug("processNewspaperEvents: shutting down")
-			return
-		}
-	}
-}
-
-func (network *Network) handleNewspaperEvent(event NewspaperEvent) {
-	// Verify signature if present
-	if event.Signature != "" {
-		// Get public key - try from event status first, then from known neighbor
-		var pubKey []byte
-		if event.Status.PublicKey != "" {
-			var err error
-			pubKey, err = ParsePublicKey(event.Status.PublicKey)
-			if err != nil {
-				logrus.Warnf("🚨 Invalid public key in newspaper from %s", event.From)
-				return
-			}
-		} else {
-			// Try to get from known neighbor
-			pubKey = network.resolvePublicKeyForNara(event.From)
-		}
-
-		if pubKey != nil && !event.Verify(pubKey) {
-			logrus.Warnf("🚨 Invalid signature on newspaper from %s, allowing for now...", event.From) // TODO(signatures)
-			//return
-		}
-	}
-
-	network.local.mu.Lock()
-	nara, present := network.Neighbourhood[event.From]
-	network.local.mu.Unlock()
-	if present {
-		nara.mu.Lock()
-		// Warn if public key changed
-		if event.Status.PublicKey != "" && nara.Status.PublicKey != "" && nara.Status.PublicKey != event.Status.PublicKey {
-			logrus.Warnf("⚠️  Public key changed for %s! Old: %s..., New: %s...",
-				event.From,
-				truncateKey(nara.Status.PublicKey),
-				truncateKey(event.Status.PublicKey))
-		}
-		// Log key field differences before updating
-		var changes []string
-		if nara.Status.Flair != event.Status.Flair && event.Status.Flair != "" {
-			changes = append(changes, fmt.Sprintf("Flair:%s→%s", nara.Status.Flair, event.Status.Flair))
-		}
-		if nara.Status.Trend != event.Status.Trend && event.Status.Trend != "" {
-			changes = append(changes, fmt.Sprintf("Trend:%s→%s", nara.Status.Trend, event.Status.Trend))
-		}
-		if nara.Status.Chattiness != event.Status.Chattiness {
-			changes = append(changes, fmt.Sprintf("Chattiness:%d→%d", nara.Status.Chattiness, event.Status.Chattiness))
-		}
-		if nara.Version != "" && event.Status.Version != "" && nara.Version != event.Status.Version {
-			changes = append(changes, fmt.Sprintf("Version:%s→%s", nara.Version, event.Status.Version))
-		}
-		if len(changes) > 0 && network.logService != nil {
-			network.logService.BatchNewspaper(event.From, strings.Join(changes, ", "))
-		}
-		nara.Status.setValuesFrom(event.Status)
-		nara.mu.Unlock()
-	} else {
-		logrus.Printf("%s posted a newspaper story (whodis?)", event.From)
-		nara = NewNara(event.From)
-		nara.Status.setValuesFrom(event.Status)
-		if network.local.Me.Status.Chattiness > 0 && !network.ReadOnly {
-			network.heyThere()
-		}
-		network.importNara(nara)
-	}
-
-	// The newspaper itself is an event emitted by them - they prove themselves
-	network.recordObservationOnlineNara(event.From, 0)
-}
-
-func (network *Network) processHowdyEvents() {
-	for {
-		select {
-		case event := <-network.howdyInbox:
-			network.handleHowdyEvent(event)
-		case <-network.ctx.Done():
-			logrus.Debug("processHowdyEvents: shutting down")
-			return
-		}
-	}
-}
-
-func (network *Network) handleHowdyEvent(howdy HowdyEvent) {
-	// Batch observed howdys for aggregated logging
-	if network.logService != nil {
-		network.logService.BatchObservedHowdy(howdy.From, howdy.To)
-	}
-
-	// Notify coordinator that we saw a howdy (for self-selection)
-	network.onHowdySeen(howdy)
-
-	// Verify signature
-	if !howdy.Verify() {
-		logrus.Debugf("howdy from %s failed signature verification", howdy.From)
-		return
-	}
-
-	// 1. Import the sender (so we know about them)
-	sender := NewNara(howdy.From)
-	sender.Status = howdy.Me
-	network.importNara(sender)
-	// The howdy itself is an event they emitted - they prove themselves
-	network.recordObservationOnlineNara(howdy.From, 0)
-
-	// 2. If this howdy is for us, process it fully
-	if howdy.To == network.meName() {
-		// Log via LogService (batched)
-		if network.logService != nil {
-			network.logService.BatchHowdyForMe(howdy.From)
-		}
-
-		// Apply their observation about us (includes StartTime!)
-		network.collectStartTimeVote(howdy.You, howdy.Me.HostStats.Uptime)
-
-		// 3. Import the neighbors they shared
-		for _, neighbor := range howdy.Neighbors {
-			n := NewNara(neighbor.Name)
-			n.Status.PublicKey = neighbor.PublicKey
-			n.Status.MeshIP = neighbor.MeshIP
-			if neighbor.MeshIP != "" {
-				n.Status.MeshEnabled = true
-			}
-			if neighbor.ID != "" {
-				n.Status.ID = neighbor.ID
-				n.ID = neighbor.ID
-			}
-			network.importNara(n)
-			// Store their observation about this neighbor
-			network.mergeExternalObservation(neighbor.Name, neighbor.Observation)
-		}
-	}
-
-	network.Buzz.increase(1)
-}
-
-// startTimeVote represents a vote for our start time from another nara
-type startTimeVote struct {
-	value  int64
-	uptime uint64
-}
-
-// collectStartTimeVote collects a vote about our start time from a howdy response
-func (network *Network) collectStartTimeVote(obs NaraObservation, senderUptime uint64) {
-	// Skip if no start time info
-	if obs.StartTime == 0 {
-		return
-	}
-
-	network.startTimeVotesMu.Lock()
-	defer network.startTimeVotesMu.Unlock()
-
-	network.startTimeVotes = append(network.startTimeVotes, startTimeVote{
-		value:  obs.StartTime,
-		uptime: senderUptime,
-	})
-
-	// Don't apply consensus early - let formOpinion() handle it with more data
-	// after boot recovery completes (~3+ minutes)
-}
-
-// applyStartTimeConsensus applies consensus to determine our start time
-// Must be called with startTimeVotesMu held
-func (network *Network) applyStartTimeConsensus() {
-	if len(network.startTimeVotes) == 0 {
-		return
-	}
-
-	// Convert to consensusValue format
-	var values []consensusValue
-	for _, vote := range network.startTimeVotes {
-		values = append(values, consensusValue{
-			value:  vote.value,
-			weight: vote.uptime,
-		})
-	}
-
-	// Use existing consensus algorithm
-	const tolerance int64 = 60 // seconds, handles clock drift
-	result := consensusByUptime(values, tolerance, false)
-
-	if result > 0 {
-		obs := network.local.getMeObservation()
-		if obs.StartTime != result {
-			obs.StartTime = result
-			network.local.setMeObservation(obs)
-			logrus.Infof("🕰️ Recovered start time via howdy consensus: %d (from %d votes)", result, len(network.startTimeVotes))
-		}
-	}
-}
-
 // mergeExternalObservation merges an external observation into our local observations
 func (network *Network) mergeExternalObservation(name string, external NaraObservation) {
 	obs := network.local.getObservation(name)
@@ -1898,432 +1205,6 @@ func (network *Network) mergeExternalObservation(name string, external NaraObser
 	}
 
 	network.local.setObservation(name, obs)
-}
-
-func (network *Network) processHeyThereEvents() {
-	for {
-		select {
-		case event := <-network.heyThereInbox:
-			network.handleHeyThereEvent(event)
-		case <-network.ctx.Done():
-			logrus.Debug("processHeyThereEvents: shutting down")
-			return
-		}
-	}
-}
-
-func (network *Network) handleHeyThereEvent(event SyncEvent) {
-	if event.Service != ServiceHeyThere || event.HeyThere == nil {
-		return
-	}
-
-	heyThere := event.HeyThere
-
-	// Verify SyncEvent signature using the public key from the payload
-	if event.IsSigned() && heyThere.PublicKey != "" {
-		pubKey, err := ParsePublicKey(heyThere.PublicKey)
-		if err != nil {
-			logrus.Warnf("🚨 Invalid public key in hey_there from %s: %v", heyThere.From, err)
-			return
-		}
-		if !event.VerifyWithKey(pubKey) {
-			logrus.Warnf("🚨 Invalid signature on hey_there from %s", heyThere.From)
-			return
-		}
-	}
-
-	// The hey_there itself is an event they emitted - they prove themselves
-	// (LogService handles logging via ledger watching)
-	network.recordObservationOnlineNara(heyThere.From, event.Timestamp/1e9)
-
-	// Add to ledger for gossip propagation
-	if network.local.SyncLedger != nil {
-		network.local.SyncLedger.AddEvent(event)
-		if network.local.Projections != nil {
-			network.local.Projections.Trigger()
-		}
-		network.broadcastSSE(event)
-	}
-
-	// Store PublicKey and MeshIP from the hey_there event
-	if heyThere.PublicKey != "" || heyThere.MeshIP != "" {
-		network.local.mu.Lock()
-		nara, present := network.Neighbourhood[heyThere.From]
-		network.local.mu.Unlock()
-
-		if present {
-			nara.mu.Lock()
-			// Warn if public key changed
-			if heyThere.PublicKey != "" && nara.Status.PublicKey != "" && nara.Status.PublicKey != heyThere.PublicKey {
-				logrus.Warnf("⚠️  Public key changed for %s! Old: %s..., New: %s...",
-					heyThere.From,
-					truncateKey(nara.Status.PublicKey),
-					truncateKey(heyThere.PublicKey))
-			}
-			if heyThere.PublicKey != "" {
-				nara.Status.PublicKey = heyThere.PublicKey
-			}
-			if heyThere.MeshIP != "" {
-				nara.Status.MeshIP = heyThere.MeshIP
-				nara.Status.MeshEnabled = true
-			}
-			if heyThere.ID != "" {
-				nara.Status.ID = heyThere.ID
-				nara.ID = heyThere.ID
-			}
-			nara.mu.Unlock()
-			logrus.Infof("📝 Updated %s: PublicKey=%s..., MeshIP=%s, ID=%s",
-				heyThere.From,
-				truncateKey(heyThere.PublicKey),
-				heyThere.MeshIP,
-				heyThere.ID)
-		}
-	}
-
-	// Start howdy coordinator with self-selection timer
-	// The coordinator uses random delays (0-3s) to spread responses and prevent thundering herd
-	if !network.ReadOnly {
-		network.startHowdyCoordinator(heyThere.From)
-	}
-	network.Buzz.increase(1)
-}
-
-// howdyCoordinator tracks howdy responses for a given hey_there
-type howdyCoordinator struct {
-	target         string          // who said hey_there
-	seen           int             // how many howdys we've seen for this target
-	mentionedNaras map[string]bool // naras already mentioned in other howdys
-	timer          *time.Timer
-	responded      bool
-	mu             sync.Mutex
-}
-
-// startHowdyCoordinator begins the self-selection process to potentially respond with howdy
-func (network *Network) startHowdyCoordinator(to string) {
-	if network.ReadOnly {
-		return
-	}
-
-	coord := &howdyCoordinator{
-		target:         to,
-		mentionedNaras: make(map[string]bool),
-	}
-	network.howdyCoordinators.Store(to, coord)
-
-	// Initial random delay: 0-3 seconds to spread howdy responses
-	// Skip jitter in tests for faster discovery
-	var delay time.Duration
-	if network.testSkipJitter {
-		delay = 0
-	} else {
-		delay = time.Duration(rand.Intn(3000)) * time.Millisecond
-	}
-	coord.timer = time.AfterFunc(delay, func() {
-		network.maybeRespondWithHowdy(coord)
-	})
-
-	// Clean up after 30 seconds
-	time.AfterFunc(30*time.Second, func() {
-		network.howdyCoordinators.Delete(to)
-	})
-}
-
-// maybeRespondWithHowdy sends a howdy if we haven't already and <10 have been sent
-func (network *Network) maybeRespondWithHowdy(coord *howdyCoordinator) {
-	coord.mu.Lock()
-	defer coord.mu.Unlock()
-
-	if coord.responded || coord.seen >= 10 {
-		return
-	}
-	coord.responded = true
-	coord.seen++
-
-	// Select neighbors to share
-	neighbors := network.selectNeighborsForHowdy(coord.target, coord.mentionedNaras)
-
-	network.local.Me.mu.Lock()
-	meStatus := network.local.Me.Status
-	network.local.Me.mu.Unlock()
-
-	event := &HowdyEvent{
-		From:      network.meName(),
-		To:        coord.target,
-		Seq:       coord.seen,
-		You:       network.local.getObservation(coord.target),
-		Neighbors: neighbors,
-		Me:        meStatus,
-	}
-	event.Sign(network.local.Keypair)
-	network.postEvent("nara/plaza/howdy", event)
-
-	logrus.Infof("👋 Sent howdy to %s (seq=%d, neighbors=%d)", coord.target, coord.seen, len(neighbors))
-}
-
-// onHowdySeen updates the coordinator when we see another nara's howdy
-func (network *Network) onHowdySeen(howdy HowdyEvent) {
-	if val, ok := network.howdyCoordinators.Load(howdy.To); ok {
-		coord := val.(*howdyCoordinator)
-		coord.mu.Lock()
-		coord.seen++
-
-		// Track which naras have been mentioned so we don't duplicate
-		coord.mentionedNaras[howdy.From] = true
-		for _, neighbor := range howdy.Neighbors {
-			coord.mentionedNaras[neighbor.Name] = true
-		}
-
-		if coord.seen >= 10 && coord.timer != nil {
-			coord.timer.Stop() // Don't bother responding
-		}
-		coord.mu.Unlock()
-	}
-}
-
-// selectNeighborsForHowdy selects up to 10 neighbors to share in a howdy response
-// Priority: online naras first (not already mentioned), then offline if needed
-func (network *Network) selectNeighborsForHowdy(exclude string, alreadyMentioned map[string]bool) []NeighborInfo {
-	maxNeighbors := 10
-	if network.isShortMemoryMode() {
-		maxNeighbors = 5
-	}
-
-	type naraWithLastSeen struct {
-		name     string
-		lastSeen int64
-		online   bool
-	}
-
-	var candidates []naraWithLastSeen
-
-	network.local.mu.Lock()
-	for _, nara := range network.Neighbourhood {
-		// Skip self, target, and already mentioned
-		if nara.Name == network.meName() || nara.Name == exclude {
-			continue
-		}
-		if alreadyMentioned[nara.Name] {
-			continue
-		}
-
-		obs := network.local.getObservationLocked(nara.Name)
-		candidates = append(candidates, naraWithLastSeen{
-			name:     nara.Name,
-			lastSeen: obs.LastSeen,
-			online:   obs.isOnline(),
-		})
-	}
-	network.local.mu.Unlock()
-
-	// Sort: online first, then by least-recently-active (oldest LastSeen first)
-	sort.Slice(candidates, func(i, j int) bool {
-		// Online naras come first
-		if candidates[i].online != candidates[j].online {
-			return candidates[i].online
-		}
-		// Then sort by LastSeen (ascending = least recently active first)
-		return candidates[i].lastSeen < candidates[j].lastSeen
-	})
-
-	// Take up to maxNeighbors
-	if len(candidates) > maxNeighbors {
-		candidates = candidates[:maxNeighbors]
-	}
-
-	// Build result
-	var result []NeighborInfo
-	network.local.mu.Lock()
-	for _, c := range candidates {
-		nara, ok := network.Neighbourhood[c.name]
-		if !ok {
-			continue
-		}
-		nara.mu.Lock()
-		info := NeighborInfo{
-			Name:        nara.Name,
-			PublicKey:   nara.Status.PublicKey,
-			MeshIP:      nara.Status.MeshIP,
-			ID:          nara.Status.ID,
-			Observation: network.local.getObservationLocked(nara.Name),
-		}
-		nara.mu.Unlock()
-		result = append(result, info)
-	}
-	network.local.mu.Unlock()
-
-	return result
-}
-
-// truncateKey returns first 8 chars of a key for logging
-func truncateKey(key string) string {
-	if len(key) > 8 {
-		return key[:8]
-	}
-	return key
-}
-
-// heyThere broadcasts identity via MQTT as a signed SyncEvent.
-// This publishes the same SyncEvent format used in gossip, ensuring
-// consistent attestation across all distribution channels.
-func (network *Network) heyThere() {
-	if network.ReadOnly {
-		return
-	}
-
-	// Always record our own online observation (needed for local state)
-	network.recordObservationOnlineNara(network.meName(), 0)
-
-	ts := int64(5) // seconds
-	if (time.Now().Unix() - network.LastHeyThere) <= ts {
-		return
-	}
-	network.LastHeyThere = time.Now().Unix()
-
-	// Create signed SyncEvent - same format for MQTT and gossip
-	event := NewHeyThereSyncEvent(
-		network.meName(),
-		network.local.Me.Status.PublicKey,
-		network.local.Me.Status.MeshIP,
-		network.local.ID,
-		network.local.Keypair,
-	)
-
-	// Publish to MQTT
-	topic := "nara/plaza/hey_there"
-	network.postEvent(topic, event)
-	logrus.Printf("%s: 👋 (MQTT)", network.meName())
-
-	// Also add to our ledger for gossip propagation and projection tracking
-	if network.local.SyncLedger != nil {
-		network.local.SyncLedger.AddEvent(event)
-		// Trigger projection so it knows we're online
-		if network.local.Projections != nil {
-			network.local.Projections.Trigger()
-		}
-	}
-
-	network.Buzz.increase(2)
-}
-
-func (network *Network) processChauEvents() {
-	for {
-		select {
-		case event := <-network.chauInbox:
-			network.handleChauEvent(event)
-		case <-network.ctx.Done():
-			logrus.Debug("processChauEvents: shutting down")
-			return
-		}
-	}
-}
-
-func (network *Network) handleChauEvent(syncEvent SyncEvent) {
-	if syncEvent.Service != ServiceChau || syncEvent.Chau == nil {
-		return
-	}
-
-	chau := syncEvent.Chau
-	if chau.From == network.meName() || chau.From == "" {
-		return
-	}
-
-	// Verify SyncEvent signature using the public key from the payload (if available)
-	// or from our neighbourhood (if we already know them)
-	if syncEvent.IsSigned() {
-		var pubKey []byte
-		if chau.PublicKey != "" {
-			var err error
-			pubKey, err = ParsePublicKey(chau.PublicKey)
-			if err != nil {
-				logrus.Warnf("⚠️  Invalid public key in chau from %s: %v", chau.From, err)
-				return
-			}
-		} else {
-			pubKey = network.resolvePublicKeyForNara(chau.From)
-		}
-		if pubKey != nil && !syncEvent.VerifyWithKey(pubKey) {
-			logrus.Warnf("⚠️  chau from %s has invalid signature, ignoring", chau.From)
-			return
-		}
-	}
-
-	network.local.mu.Lock()
-	existingNara, present := network.Neighbourhood[chau.From]
-	network.local.mu.Unlock()
-
-	// Check for public key changes
-	if present && existingNara.Status.PublicKey != "" && chau.PublicKey != "" {
-		if existingNara.Status.PublicKey != chau.PublicKey {
-			logrus.Warnf("⚠️  PUBLIC KEY CHANGED for %s! old=%s new=%s",
-				chau.From, truncateKey(existingNara.Status.PublicKey), truncateKey(chau.PublicKey))
-		}
-	}
-
-	// Update the nara's public key and ID if provided
-	if present && chau.PublicKey != "" {
-		existingNara.Status.PublicKey = chau.PublicKey
-	}
-	if present && chau.ID != "" {
-		existingNara.Status.ID = chau.ID
-		existingNara.ID = chau.ID
-	}
-
-	// Add to ledger for gossip propagation
-	if network.local.SyncLedger != nil {
-		network.local.SyncLedger.AddEvent(syncEvent)
-		if network.local.Projections != nil {
-			network.local.Projections.Trigger()
-		}
-		network.broadcastSSE(syncEvent)
-	}
-
-	observation := network.local.getObservation(chau.From)
-	previousState := observation.Online
-	observation.Online = "OFFLINE"
-	observation.LastSeen = time.Now().Unix()
-	network.local.setObservation(chau.From, observation)
-
-	// Record offline observation event if state changed
-	if previousState == "ONLINE" && !network.local.isBooting() && network.local.SyncLedger != nil {
-		obsEvent := NewObservationSocialSyncEvent(network.meName(), chau.From, ReasonOffline, network.local.Keypair)
-		network.local.SyncLedger.AddSocialEventFiltered(obsEvent, network.local.Me.Status.Personality)
-		// LogService handles logging via ledger watching
-	}
-
-	network.Buzz.increase(2)
-}
-
-func (network *Network) Chau() {
-	if network.ReadOnly {
-		return
-	}
-
-	// Update our own observation to OFFLINE
-	observation := network.local.getMeObservation()
-	observation.Online = "OFFLINE"
-	observation.LastSeen = time.Now().Unix()
-	network.local.setMeObservation(observation)
-
-	// Create signed SyncEvent - same format for MQTT and gossip
-	event := NewChauSyncEvent(
-		network.meName(),
-		network.local.Me.Status.PublicKey,
-		network.local.ID,
-		network.local.Keypair,
-	)
-
-	// Publish to MQTT
-	topic := "nara/plaza/chau"
-	network.postEvent(topic, event)
-	// LogService handles logging via ledger watching
-
-	// Also add to ledger for gossip propagation
-	if network.local.SyncLedger != nil {
-		network.local.SyncLedger.AddEvent(event)
-		if network.local.Projections != nil {
-			network.local.Projections.Trigger()
-		}
-	}
 }
 
 func (network *Network) oldestNaraBarrio() Nara {
@@ -4230,95 +3111,6 @@ func (network *Network) isShortMemoryMode() bool {
 		return false
 	}
 	return network.local.MemoryProfile.Mode == MemoryModeShort
-}
-
-func (network *Network) recoverSelfStartTimeFromMesh() {
-	if network.local.SyncLedger == nil || network.tsnetMesh == nil || network.local.isBooting() {
-		return
-	}
-
-	obs := network.local.getMeObservation()
-	if obs.StartTime > 0 {
-		return
-	}
-
-	online := network.NeighbourhoodOnlineNames()
-	if len(online) == 0 {
-		return
-	}
-
-	// Ask a few neighbors for their opinions about our start time.
-	rand.Shuffle(len(online), func(i, j int) {
-		online[i], online[j] = online[j], online[i]
-	})
-
-	targetCount := len(online)
-	if targetCount > 3 {
-		targetCount = 3
-	}
-
-	client := network.getMeshHTTPClient()
-	if client == nil {
-		return
-	}
-	subjects := []string{network.meName()}
-	totalAdded := 0
-
-	for _, neighbor := range online[:targetCount] {
-		ip := network.getMeshIPForNara(neighbor)
-		if ip == "" {
-			continue
-		}
-
-		events, respVerified := network.fetchSyncEventsFromMesh(client, ip, neighbor, subjects, 0, 1, 500)
-		if len(events) == 0 {
-			continue
-		}
-
-		filtered := make([]SyncEvent, 0, len(events))
-		for _, event := range events {
-			if event.Service != ServiceObservation || event.Observation == nil {
-				continue
-			}
-			if event.Observation.Subject != network.meName() {
-				continue
-			}
-			filtered = append(filtered, event)
-		}
-
-		if len(filtered) == 0 {
-			continue
-		}
-
-		added, warned := network.MergeSyncEventsWithVerification(filtered)
-		totalAdded += added
-		verifiedStr := ""
-		if respVerified && warned == 0 {
-			verifiedStr = " ✓"
-		} else if warned > 0 {
-			verifiedStr = fmt.Sprintf(" ⚠%d", warned)
-		}
-		logrus.Printf("📦 start time recovery from %s: received %d events, merged %d%s", neighbor, len(filtered), added, verifiedStr)
-	}
-
-	if totalAdded == 0 {
-		return
-	}
-
-	// Try to recover start time from event-based consensus
-	if network.local.Projections != nil {
-		updated := network.local.getMeObservation()
-		before := updated.StartTime
-		network.local.Projections.Opinion().RunOnce()
-		opinion := network.local.Projections.Opinion().DeriveOpinionWithValidation(network.meName())
-		if updated.StartTime == 0 && opinion.StartTime > 0 {
-			updated.StartTime = opinion.StartTime
-		}
-		if updated.StartTime > 0 && updated.StartTime != before {
-			network.local.setObservation(network.meName(), updated)
-			logrus.Printf("🕰️ recovered start time for %s via event consensus: %d", network.meName(), updated.StartTime)
-		}
-	}
 }
 
 // performBackgroundSyncViaMesh performs background sync with a specific neighbor via mesh
