+++ presence_projection.go
